reorganized git package
